+++ libs/libvga/pos/putstrpos.c
@@ -21,10 +21,11 @@
 #include <libvga.h>
 #include <string.h>
 #include <libssp.h>
+#include <system/types.h>
 
 
 void putstrpos(const char *string, unsigned char x, unsigned char y, unsigned char fg_color, unsigned char bg_color, int start) {
-  for (int i = 0; i < strlen(string); i++) {
+  for (size_t i = 0; i < strlen(string); i++) {
 	if (string[i] == '\n') {
 	  x = start;
 	  y++;
